renamed 'typescript' folder to 'src'
