+++ server/src/Scoreboard.ts
